requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1

# Web Scraping & Crawling
scrapy==2.11.0
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# Data Processing
pandas==2.1.4
numpy==1.25.2
elasticsearch==8.11.0

# Website Checking
dnspython==2.4.2
python-whois==0.8.0
urllib3==2.1.0

# Task Queue
celery==5.3.4
flower==2.0.1

# Geospatial
geopy==2.4.1

# API & HTTP
httpx==0.25.2
aiohttp==3.9.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Environment & Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Export & Reporting
openpyxl==3.1.2
xlsxwriter==3.1.9

# Machine Learning (Future)
scikit-learn==1.3.2
nltk==3.8.1 